Include Pydantic models in the definition parsing tests.
